refactor: update CharacterSheet import to use named export from index file
